+++ intelmq/tests/bots/collectors/mail/test_collector_attach.py
@@ -8,30 +8,9 @@ import unittest.mock as mock
 
 import intelmq.lib.test as test
 
-if os.getenv('INTELMQ_TEST_EXOTIC'):
-    from imbox.parser import parse_email
-    with open(os.path.join(os.path.dirname(__file__), 'foobarzip.eml')) as handle:
-        EMAIL_FOOBAR = parse_email(handle.read())
-
-    class MockedImbox():
-        _connected = False
-
-        def __init__(self, hostname, username=None, password=None, ssl=True,
-                     port=None, ssl_context=None, policy=None, starttls=False):
-            pass
-
-        def messages(self, *args, **kwargs):
-            yield 0, EMAIL_FOOBAR
-
-        def mark_seen(self, uid):
-            pass
-
-        def logout(self):
-            pass
-
-
 from intelmq.bots.collectors.mail.collector_mail_attach import MailAttachCollectorBot
-
+if os.getenv('INTELMQ_TEST_EXOTIC'):
+    from .lib import MockedZipImbox
 
 REPORT_FOOBARZIP = {
                     '__type': 'Report',
@@ -63,7 +42,7 @@ class TestMailAttachCollectorBot(test.BotTestCase, unittest.TestCase):
                          }
 
     def test_one(self):
-        with mock.patch('imbox.Imbox', new=MockedImbox):
+        with mock.patch('imbox.Imbox', new=MockedZipImbox):
             self.run_bot()
         self.assertMessageEqual(0, REPORT_FOOBARZIP)